Working on adding TMP dropdown
1. Added TMP dropdown to scene 1.2
2. Added TMP assets
3. Made modifications to TMP dropdown to support the creation of lattices
4. Almost everything works, but the dropdown text is not centered

+++ Assets/MobileIonScript.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class MobileIonScript : MonoBehaviour
 {
-    public int charge;
+    public float charge;
     public bool ChargeActive;
     private Vector3 velocity;
     private float temp;
